Fix MultiEnvTestFilter to pass inner test classes (#5775)
Previously the filter required multi-version test extensions to
recognise all test classes in their `accepts` method. This proved
difficult for inner test classes that generally do not carry the
`@ExtendWith` annotation.

Moreover, the old filtering logic was overzealous in requiring
extensions to explicitly "accept" test classes, which would generally
be annotated with the extension anyway.

This change removes the `accepts` method from `MultiEnvTestExtension`
and simplifies the filtering logic to be based only on the presence
of appropriate extension annotations.

MultiEnvExtensionRegistry is enhanced to follow class nesting chains
to find extension annotations.