Validate runtime parameters in service implementations.
This is to follow up on #5740, which introduced validation
annotations to Nessie backend service interfaces, but did not enforce
them.

This change introduced proper CDI beans for backend services in order
for Quarkus and Weld to automatically validate method parameters in
runtime.

Quarkus does the right validation out of the box.

Weld needs an explicit dependency on a CDI validator lib, for which
the Hibernate Validator is used. The `@ValidateOnExecution`
annotations are also required for this use case.

Namespace tests are adjusted to provide proper expected hashes during
commit operations. These tests used to work fine over API v1, which
is lenient in its NamespaceApi implementation, but using strict
expected hashes is less ambiguous, matches old test behaviour in API
v1, and works in both API versions.