Update V2 commit-response to include generated content-IDs
This changes `org.projectnessie.model.CommitResponse` to include the content-IDs of the _newly_ added `Content` objects - for REST API v2.

The Nessie API requires people to _not_ set the `Content.id` for _new_ content (think: `CREATE TABLE`), Nessie will generate the content-ID in that case. This behavior will be strictly enforced with #5641. So this change adds a new `List<AddedContent>` to `CommitResponse`, which contains the generated content-IDs, so use cases that need the content-ID can just inspect the `CommitResponse` and do not have to issue a `getContents()` to get the generated content-ID.