Fix Jacoco reporting in Nessie Quarkus projects
Jacoco report generation via Quarkus is not so great, there were a bunch of issues, some breaking the build (flaky)

This change disables Jacoco report generation by Quarkus and changes it to let the Jacoco report task generate the reports.

Also update the GH workflows to run `codeCoverageReport` with `intTest`.