Separate version catalog for `buildSrc/`
Have a separate version catalog into one for `buildSrc/`.

This helps Gradle to not invalidate the build cache for _every_ dependency update.

Background: `buildSrc/` used the `libs` version catalog, but if that one is updated, the cached build artifacts for `buildSrc/` are invalidated, which then in turn causes _all_ projects' build caches to be invalidated.