Make `QuarkusBuild` cacheable
The output of the `QuarkusBuild` task is, by default, not cacheable. This means, that if the output does not exist locally, the task will run. But OTOH the output of `QuarkusBuild` is not repeatable, the contents of the jars and other files generated by Quarkus vary from build to build w/ the exact same inputs, which causes dependent tasks, which are all the `intTest` tasks, that use the Nessie Quarkus Runner, will re-run, because their input (the output of the `QuarkusBuild`) changed.

There are a bunch of projects that use the Nessie Quarkus Runner, and those run for quite a while (10 minutes at least in CI).

So caching the outputs of `QuarkusBuild` should at least be tried out on CI. Although the cache entries for `QuarkusBuild` are quite large (>110M), it probably justifies the build time improvements.

Side note: this affects only usages of the Nessie Quarkus Runner in the Nessie source tree. Projects that use _released_ Nessie versions are not affected.