Produce NessieBadRequestException when paging is not available
Affected APIs:
* getAllReferences()
* getEntries()
* getDiff()

Previously, if the page has to be cut short in the responses from
these APIs (e.g. if `maxRecords` is set). The response would have
`hasMore == true` but `pageToken == null`. This can be confusing to
clients.

With this change the service with cause a `NessieBadRequestException`
when it has to limit the response size, but cannot generate a paging
token to clearly indicate that paging is not supported in these
situations.

Clients may still observe that requests with a large `maxRecords`
value succeed, but with a smaller `maxRecords` value fail. This
is because the decision to limit the page size is made at the REST
API level, but the token is generated at the storage level and exact
behaviour depends on the available data.

Full pagination support is to be added with the new model (#5641).