Fix dependencies in :nessie-compatibility-common
"implementation" dependency on In-Memory and RocksDB adapters is
required because `OlderNessieClientsExtension` and
`NessieUpgradesExtension` load them dynamically in runtime.

Switch main artifact from :nessie-jaxrs to :nessie-compatibility-common
after 0.48.1 because dependencies were rearranged. 0.47.1 itself is not
really usable with the "compatibility" test framework, but adding
workarounds for it does not look worth the while. The next version
should be usable.