Make Nessie build with `javax.*` and `jakarta.*` dependencies
This change adds the Jakarta API dependencies next to the (old) javax dependencies. Hint: the findbugs JSR-305 dependency is the "javax-equivalent" for jakarta-annotations.

There's no production code change in this PR, but dependency configurations/scopes have changed. Assuming that server _implementations_ come with the "right" dependencies for the annotations, which should work fine.

This change alone will _not_ make Nessie work with Jakarte EE 10, because
* neither the javax-annotations are not "duplicated" with their "jakarta counterparts",
* nor are the javax-interface implementations duplicated with their "jakarta counterparts" (think: `implements javax.ws.rs.ext.ParamConverterProvider` or `implements javax.ws.rs.ext.ExceptionMapper`).