Gatling simulation for a mixed workload (#6068)
Current Gatling simulations target only one specific use case and are
always focued on "commit rate".

The simulation in this PR supports a mixed workload.

"Workflow" of the whole simulation
1. Setup phase:
1. Create a starting branch with many Iceberg tables, allows creating
multiple tables in a single commit to reduce the setup time.
2. Create additional branches, if more than one branch is configured,
based on the HEAD of the branch from step 1.1.
2. "Mixed workload" phase with three use cases. All can be tweaked with
the number of simulated users and the rate per second for each of the
simulated users.
1. UI user: 1. List all tables 2. Get current content of one tables 3.
Update that table (Nessie commit)
    2. "Readers" (read from N tables at some fixed rate):
        1. Get current content of one or more tables
2. "Writer" (think: a scheduled job committing at some fixed rate): 1.
Get current content of one tables 1. Update that table (Nessie commit)

The "global" set of tables consists of a "pool" of tables and a subset
of that "pool" as "active" tables to simulate the (probably common)
scenario that people have many 10000s tables, but only actively use a
fraction of those.