Fail test discovery in MultiEnvTestEngine if no environments are defined (#6108)
This is to prevent silently ignoring tests in case of configuration
mistakes like forgetting to set the `nessie.versions` system property or
a `META-INF/nessie-compatibility.properties` file.

The default is to fail during test discovery. The failure can be
suppressed, when necessary, by setting the java system property
`org.projectnessie.junit.engine.ignore-empty-environments` to "true".