Allow all GH WFs to become "required checks" (#6131)
A couple of PR workflows are "optional", but if those are enabled, those
should succeed, especially for automatically merged PRs like those
opened by renovate.

A PR WF that's not run is marked as "Skipped". For the "required checks"
a skipped check means "failed". The [workaround described
here](https://docs.github.com/en/repositories/configuring-branches-and-merges-in-your-repository/defining-the-mergeability-of-pull-requests/troubleshooting-required-status-checks#handling-skipped-but-required-checks)
recommends adding another WF with the same WF and step names as the
"skipped one", see the description in the link for details.

The proposed workaround is not really nice, WFs/checks appear twice in
the list of PR checks and WFs _appear_ as exact duplicates in the list
of workflows (in the repository's "Actions" tab).

This change refactors the "global" `if:` of the four workflows to become
an environment variable, which is checked for every step. Although that
adds some boilerplate to the WF definition, the overall appearance is
much nicer.

This change allows making the jobs in these four WFs to become "required
status checks".