Remove rocksdb.dbPath from Helm chart values (#6149)
This commit removes the rocksdb.dbPath setting from the values.yaml file of Nessie's chart.

While this setting is exposed to users running Nessie outside of Kubernetes, id doesn't make a lot of sense in the Kubernetes world. What really needs to be exposed to Kubernetes users is the persistent volume claim, not the mount path on the Nessie container.