Make V2 ref path-param parsing independent of `Reference` (#6224)
`Reference` is strongly types, it can only be exactly one of the types `Branch`, `Tag` or `Detached`. When parsing the path parameter in V2 API, the actual reference type (branch or tag or detached) is usually unknown, it _can_ be typed by the called by passing the _optional_ reference type.

This change introduces a new V2-internal tuple `ParsedReference` to allow parsing a reference path parameter without knowing the reference type. The backend service interface (`TreeService`) and implementation (`TreeApiImpl`) already allows the reference-type to be optional. A follow-up ticket might refactor the `TreeService` functions to use the new `ParsedReference` type, possibly renaming it and handling the various validation rules.

Updates `TestAuthorizationRules` and adopts the authz check for delete-reference to work like the other functions, which prefer throwing a NessieReferenceNotFoundException over NessieForbiddenException. This behavior (not-found over forbidden) is undefined though - we might change it in the future, but it would require updating the authz check interface and contract as well, because the reference-type would become optional).

Fixes #6223