Ignore jEnv .java-version file
This commit adds one entry to the .gitignore file: .java-version.

This file is created by the Java version manager jEnv:

https://github.com/jenv/jenv

The file is produced when we force a different "local" Java version
than the configured global one, e.g. with the following command:

jenv local 11

This is useful when a project needs a special Java version.