GH WF: Remove `concurrency` from PR jobs
A couple PR runs indicated that some of the "required checks" did not run. The reason is that in fact the checks (WF/job runs) worked, but due the auto-labeler job re-triggered the WF runs, and the re-triggered runs were skipped due to the configured `concurrency` settings.

Sadly there's no way to have finer grained (label based) filters and concurrency settings. It might be worth to refactor the jobs, but that's out of scope of this PR.

At this point it's probably okay to bite the bullet and remove the concurrency setting to get proper PR checks back.

Example runs:
* https://github.com/projectnessie/nessie/actions/runs/4441749920
* https://github.com/projectnessie/nessie/actions/runs/4441750040
* https://github.com/projectnessie/nessie/actions/runs/4441750059