New storage: fix writetamp collision in integration tests
We must ensure that the system clock advances a little, so that C*'s next write-timestamp does not collide with the write-timestamps from an "erase repository" operation.  In C*, if a DELETE and another INSERT/UPDATE have the same write-timestamp, the DELETE wins. This makes Nessie tests fail on machines that are "fast enough".