Build/publishing: Proper Gradle module metadata for shadow-jars (#6361)
The "build tools integrations test" failed for PR #6355 for the "external Gradle build". Reason is that the relocated protobuf jar is published via the Shadow plugin. `PublishingHelperPlugin` used the recommended approach to call `ShadowExtension.component(MavenPublication)`, which is fine for Maven poms, but is not enough for Gradle module metadata (the `.module` files published along the `.pom` files).

The published Gradle module metadata files reference the "original" jar (named `*-raw.jar` in the Nessie build) in the included `apiElements` and `runtimeElements` variants, but that's wrong for Nessie and inconsistent with the published pom, because we want to _replace_ the "original" jar with the one produced by `ShadowJar`.

Relevant pieces that assemble the Gradle component to be published are taken from Gradle's `Java(Base)Plugin` and `ShadowExtension`.

The issue can be inspected in the published `.module` file (or one generated via the `generateMetadataFileForMavenPublication`, e.g. from `:nessie-spark-extensions-3.1_2.12`) project. Before this change, the file `build/publications/maven/module.json` contains these variants (simplified for readability):
```
{
  ...
  "variants": [
    {
      "name": "apiElements",
      "attributes": {
        "org.gradle.category": "library",
        "org.gradle.dependency.bundling": "external",
        "org.gradle.jvm.version": 8,
        "org.gradle.libraryelements": "jar",
        "org.gradle.usage": "java-api"
      },
      "files": [
        {
          "name": "nessie-protobuf-relocated-0.52.4-SNAPSHOT-raw.jar",
          ...
        }
      ]
    },
    ... (similar for runtimeElements)
    ... (javadoc + sources follow)
    {
      "name": "shadowRuntimeElements",
      "attributes": {
        "org.gradle.category": "library",
        "org.gradle.dependency.bundling": "shadowed",
        "org.gradle.libraryelements": "jar",
        "org.gradle.usage": "java-runtime"
      },
      "files": [
        {
          "name": "nessie-protobuf-relocated-0.52.4-SNAPSHOT.jar",
          ...
        }
      ]
    }
  ]
}
```
With this change it looks as follows, note that the the duplicated `java-runtime` usage is gone and a `java-api` for the the shadowed bundling is present. This way, "external" builds can properly select the required variant.
```
{
  ...
  "variants": [
    {
      "name": "shadowApiElements",
      "attributes": {
        "org.gradle.category": "library",
        "org.gradle.dependency.bundling": "shadowed",
        "org.gradle.jvm.version": 8,
        "org.gradle.libraryelements": "jar",
        "org.gradle.usage": "java-api"
      },
      "files": [
        {
          "name": "nessie-protobuf-relocated-0.52.4-SNAPSHOT.jar",
          ...
        }
      ]
    },
    {
      "name": "shadowRuntimeElements",
      "attributes": {
        "org.gradle.category": "library",
        "org.gradle.dependency.bundling": "shadowed",
        "org.gradle.libraryelements": "jar",
        "org.gradle.usage": "java-runtime"
      },
      "files": [
        {
          "name": "nessie-protobuf-relocated-0.52.4-SNAPSHOT.jar",
          ...
        }
      ]
    },
    ... (javadoc + sources follow)
  ]
}
```