Persist: replace `updateObj()` with `upsertObj()`
The only purpose of `updateObj()` is to allow the Nessie repository
import to enhance the freshly imported `CommitObj`s, which are marked
with `incompleteIndex==true`, with properly populated incremental and
reference indexes.

During the import operation, we do not really care, whether the object
to be updated (the commit) does not exist, because that will not be the
case, because the import already stored all the commits. TL;DR the
behavior to check whether the `Obj` passed to `updateObj()` already
exists is rather meaningless.

Using "upsert" semantics has the big benefit that those can be batched,
which is not only nice but a big performance improvement during import.