Remove old content-attachments approach
Cleanup in preparation for metadata-in-Nessie. The old "content
attachment" approach, only implemented by the old storage model, was
never really used.

This is basically the "whole spiel" to remove the content-attachment
(de)serialization code from StoreWorker, where mostly just some
parameters are removed, ending in a massive removal of code (and tests).