Let errorprone not check code generated by APTs
Errorprone fails for a recent change for #6480, complaining about an unnecessary null-check in generated immutables code (something like `int x; Map.put(Objects.requireNonNull(x), ...)`). The check itself is fine, however the code is generated, so there's not much we can do.

This change just adds the temporary directory in which errorprone works.