Enhance `NessieError` with error details
Adds a new, generic property `errorDetails` to return details about a
particular error to clients.

The field uses a generic type (in Java it's Jackson's `JsonNode`) so
old clients are able to safely ignore future enhancements like new error
details types or added properties. Parsing of `NessieError` is now more
"relaxed", unknown properties are ignored.

Clients must not assume that `errorDetails` will always be present,
because older servers may not return it.

Servers must only include the new `errorDetails` property, if the client
included the HTTP header `Nessie-Client-Spec` with an integer value >=
2.

This change also introduces `ReferenceConflicts` as the first use case
for `errorDetails`. It contains details like the particular conflict
type and related key and a mandatory message.

Fixes #4310
Part of #5645
Related to #6447