CI: Remove `CI Success` job
The idea for the `CI Success` job was to have only one required status
check, without the need to configure each individual job in `ci.yml`.
But sadly, a _skipped_ job counts as _passed_ for required checks. This
is different from `needs: [jobs]` for a job, which requires all _needed_
jobs to run and succeed, so a skipped job in `neeeds:` counts as
"failed".