Adjust testNotFoundUrls parameters
This is a follow-up to #6584

Use only truly unmapped paths for this test.

`contents/not-there-using` URLs are mapped in API v1 as requests for
the key `not-there-using`.

`trees/not-there-using` URLs are mapped in API v2 as requests for the
reference named `not-there-using`.

Even though the above two case do produce 404 response, they do not
validate the code changed in `NessieExceptionMapper` for #6569.

Also the contents request lacks some required parameters (ref name).