Allow all commit attributes for namespace operations
Adds a new optional property `commitMeta` for the create/update/delete namespace client API operations.
Falls back to the current behavior, if `commitMeta` is not specified.
Uses the current message, if the provided `commitMeta`'s `message` is empty.

Implemented for the V2 Java client, not for the V1 client.

Fixes #5615