Build: allow javac 'release' option, server modules built for Java 11
Replaces the `nessie-conventions.gradle.kts` script with a bunch of specializations for:
* "client" client facing code, for Java 8
* "spark" Nessie Spark code, client facing, Java 8
* "server" Nessie server components, Java 11
* "quarkus" Nessie components using Quarkus, Java 11
* "iceberg" Nessie components _using_ Iceberg, Java 11
* "scala" Nessie Scala code, Java 11 (perftest modules)
* "root" Nessie root projects

Moves the build logic from `buildSrc` to an included build `build-logic`. Included builds behave different than `buildSrc/`, those are "proper" projects. `buildSrc/` has the disadvantage that changing anything invalidates _all_ build caches, which is not the case for included builds.

Also removes a bunch of Kotlin files, moving the build code to the (specialized) call sites, except for `Testing.kt`, which is quite complex.

Refactored the "rather pure" Kotlin files to "precompiled script plugins", which are easier to read and understand, because those are very similar to "normal" build scripts.