Cleanup V2 MergeResponse (#6747)
Ensure that all non-V2 attributes in `MergeResponse` (those that are
present for V1, but deprecated) are properly annotated with
`@JsonView(Views.V1.class)`.

On top of the already deprecated attributes,
`MergeResponse.ContentKeyDetails.getConflictType()` is deprecated and
V1-only now as well. It already has a better replacement with
`MergeResponse.getConflictType.getConflict()`. The corresponding enum is
also deprecated now.

Fixes #5689