Allow exporting only the latest contents from a branch.
This can be useful when migrating to the new Nessie data model in cases
where the full previous change history do not have to be preserved.

Exporting only the latest state of contents can save time and storage
(smaller export data files), when the number of content objects is
significantly smaller than the number of commits.

Both old and new data models can be exported with `--only-contents-from-branch`,
even though the primary use case is exporting from the old model and importing
into the new model.