Remove Gatling runs in CI
Gatling runs are effectively not inspected or monitored, so it is wasted CI effort at the moment.

Also realized that the job that pushes the Gatling results likely annotated the wrong commits, at least on the main branch, but likely also in PRs. This happens|ed when a CI run was running and another commit was pushed - and the newly pushed commit got the Gatling results of the (older) run.