Fix `MergeKeyBehavior.DROP` in new storage model for merge
The (legacy) behavior of `MergeKeyBehavior.DROP` is to completely ignore such content keys from the source branch to be merged. The implementation in the new model wrongly implemented it to ignore the value from the source branch, but only in case of a conflict. It must also ignore such content-keys without a conflict.

This was discovered while porting the compatibility-tests to the new storage model.

This PR adds test cases to validate the behavior (and updates an existing test).