Fix reference-handling concurrency annoyances
In a scenario when references are created/deleted and concurrently listed, so that reference-creation/deletion-recovery kicks in, there are a few scenarios which can cause unnecessary but rather annoying user-facing errors (exceptions).

This has been exposed (and later verified) using the new test `AbstractReferenceLogicTests.refCreationDeletionWithConcurrentRefsListing()` (ran with up to 2000 repetitions locally, committed in this PR with 2 repetitions).