Ensure that "already existing content" does not cause conflicts
When a (merge/transplant) commit wants to "add" the exact same content
value that already exists on the target branch, that should not result
in a conflict.

Also adopt merge test cases to new-model VS tests

Let functionality to return the merge-base/common-ancestor return the `ObjId` because the caller has to perform a "load" anyway.

Some minor changes