Enhance references with created-timestamp + extended-info-obj-pointer
The intent of this change is to provide the base to fix an annoying race
where users can failures for reference creations or deletions due to a
potential race with concurrent reference queries, which trigger
reference recovery, which can legitly finish a
half-reference-creation/deletion, in these cases the original reference
creation/deletion.

This race is legit and intended, the errors should not happen though,
and can be prevented by having some form of "token" - in this case the
reference creation timestamp, which this change duplicates in the `refs`
table.

Since this change already changes all the references persistence, it
also adds a new `extendedInfoObj` attribute to references. The intent
here is to remove the (not yet) ambiguous pointer to "commt pointers"
(extended tags via `TagObj`, not exposed as a functionality yet). The
goal here is to only store the HEAD commit ID in `Reference.pointer()`
and if a reference has more information (for example that `TagObj`),
store the pointer to that one in `extendedInfoObj`. This also saves one
additional load for `CommitObjReference` implementations (like
`TagObj`) and to remove that interface.