Identify merge-base instead of common-ancestor (#7035)
Enhances the existing logic to identify the common ancestor to respect merge parents.

Externalizes the logic into a new class `MergeBase`, which combines the logic for the common-ancestor and merge-base use cases.

Currently the logic for a merge operation is to identify the common ancestor and apply the diff between the common ancestor and the "from HEAD" onto the "target HEAD".

This change will use the "best merge base" instead of the "common ancestor".

The ability to perform "merges with individual commits" and "squashed transplants" has been removed. This behavior is the only one available in Nessie V2 API.

The change ensures that "already existing content" does not cause conflicts. When a (merge/transplant) commit wants to "add" the exact same content value that already exists on the target branch, that should not result in a conflict.

Fixes #4562