Fix inability to delete some references
The new storage model's commit logic, which provides support for name indexes, is also used to provide the index of references. When a reference is about to be deleted, the logic checks whether the reference to be deleted is in that index. Unfortunately, `ReferenceLogicImpl.commitDeleteReference()` only considered the _incremental_ index and not the "spilled out" reference stripes, so that a reference that is not in the incremental index cannot be deleted. The fix is a simple one-line change to load the whole index and not just the incremental one.