Add `MergeResponseInspector` functionality to Java API
Provides a simple API to collect information that is necessary for content-aware merges.

The base for that information is already available via the `Conflict`s returned in a `MergeResponse`. The approach is based on information from the merge-request and merge-response. It collects the conflicting contents from the _merge-base_ using a Nessie-API "get-contents" call and then uses Nessie's diff-operations to identify and filter the conflicting contents from the diffs of _merge-base_-to-_merge-source_ and _merge-base_-to-_merge-target_. The content information is aggregated ("grouped" by content-ID) and returned as a Java `Stream` via the introduced API.

This change does not implement any content-aware merge operation.

This is a pure Nessie-Java-API/client change, no REST API change.