API + VersionStore changes
Adds functions to create/update repository configurations and to retrieve one or more repository config objects.

Repository config types are handles analoguous to content types. There's a registry and a way to deserialize generic repository config types. It is not possible to store repository config types that are not known to the Nessie server, this allows the Nessie server to validate the schema of the repository config object and, if necessary in the future, redact sensitive information.