Introduce `CHANGELOG.md`, include in release information
The information about a specific Nessie release is pretty rudimentary, consists only of a list of commits. Interesting, important and relevant information, for example about upgrade notes, release highlights, potentially breaking changes, is missing in the release notes.

This change introduces the `CHANGELOG.md` file. Relevant information is to be added to an appropriate section for `Unreleased`. The list of sections is not fixed and can be changed and extended, depending on the impact of an individual change/PR.

The Nessie release workflows leverage Jetbrains' [_Gradle Changelog Plugin_](https://github.com/JetBrains/gradle-changelog-plugin) to promote `Unreleased` to a release in `CHANGELOG.md` and to extract the changelog for the current release, which is then added to `site/docs/try/releases.md` and the Github release notes. The changelog-plugin is "clever enough" to purge empty sections for a release.

This is the command to extract information about the latest release
```bash
./gradlew --quiet --console=plain getChangelog --no-header --no-links
```

Using `./gradlew getChangelog --unreleased` shows information about the not-yet-released code.