Fix openapi self-reference and type ambiguity
The generated openapi-yaml has an error, because the `Type_V1` and `Type_v2` schemas have a "self-reference". This change fixes this by adding an explicit `ref` to `EntriesResponse.Entry.getType()`.

Also updates the repository-configs with the "v2" tag.

Also (re)names the type schema for `RepositoryConfig.Type` to `RepositoryConfigType`, which was ambiguous with `Content.Type`, as both resolved to the same OpenAPI schema `Type`.

Also adds an integration test to lint/validate the generated `openapi.yaml` file.

Self-referencing schemas:
```yaml
components:
  schemas:
    ...
    Type:
      description: "Declares the type of a Nessie content object, which is currently\
        \ one of ICEBERG_TABLE, DELTA_LAKE_TABLE, ICEBERG_VIEW, NAMESPACE or UDF,\
        \ which are the discriminator mapping values of the 'Content' type."
      type: string
    Type_V1:
      $ref: '#/components/schemas/Type_V1'
    Type_V2:
      $ref: '#/components/schemas/Type_V2'
```