Fix Quarkus Swagger UI (#7281)
PR #7266 broke the Swagger UI in Nessie-Quarkus. The Swagger UI in Quarkus is a bit special, and the OpenAPI spec is heavily adopted by Quarkus itself - it especially rewrites the endpoint paths. For example `/v2/config` becomes `/api/v2/config` in Quarkus' generated OpenAPI spec. The server settings in Quarkus' Swagger UI is "a bit opinionated", because it doesn't really allow to change the "base path" - and because the "base path" of the first server in the openapi.yaml file has `/api/` as the "base path", the Quarkus Swagger UI wants to invoke the endpoints at for example `/api/api/v2/config`, which doesn't work. This change fixes this.

Sadly, the Swagger UI doesn't show long descriptions for server entries - the description is appended to the single-line select-box. And descriptions for the individual URL variables are not shown at all.

Beside the above changes, this change also:
* Fixes the `from` in the `pullOpenApiSpec` task definition, so Gradle picks up changes to the source file correctly.
* Updates the build script to reference the `quarkusAppPartsBuild` task instead of `quarkusBuild` (change in the Quarkus 3 Gradle plugin).
* Fix the version number shown in Quarkus Swagger UI
* Remove the security section from Quarkus' generated openapi spec (it's wrong, points to 172.255.0.0).