Correct handling of reference-type query parameter (#7282)
Nessie API V2 uses the Java enum `ReferenceType` directly as the type of the `type` query parameter, and not a "bare `String`". This causes some special handling: the `type` parameter _must_ be present and have a corresponding value in the `ReferenceType` enum, otherwise the HTTP response will be a HTTP/404 instead of the expected HTTP/400. The query parameter could be optional using `Optional<ReferenceType>`, but that still restricts the parameter to the enum values, returning a HTTP/404.

This change refactors the `type` parameter for create-reference, assign-reference and delete-reference to be a "bare `String`" and uses a helper function to parse the value. Parameter _value_ validation is delegated to JAX-RS using a validation regex, returning a meaningful validation message.

Fixes #7229