Fix NesQuEIT IntelliJ import
The NesQuEIT project includes the Nessie sources as two Gradle builds - one for everything except Iceberg and one for the rest that has dependencies to Iceberg.

The logic in `nessie-conventions-root.gradle.kts` refers to just `projectDir`, which points to `nessie-iceberg/` instead of the Nessie source root dir.

This change fixes the above issue.