Remove call to Persist.erase in ImportPersistV1.prepareRepository
This call is redundant when using quarkus-cli that already
has a "-e" switch to erase the repository. As a consequence,
the repo was being deleted twice.

Since erase the repo is a slow operation that incurs in
a table scan, removing this call also improves the overall
important performance significantly.