GC: Fix duplicate-key error w/ PostgreSQL
As reported in #7408, Nessie GC does not work correctly against PostgreSQL.

Although there is at least one test case that exercises the relevant code paths with the "failing" data scenarios, the tests did not fail due to a (now) obvious difference between H2 (used in the tests) and PostgreSQL. This change keeps the H2 in-memory tests and introduces ITs for PostgreSQL. The "difference" was actually H2 _not_ operating in PostgreSQL-mode.

`JdbcPersisxtenceSpi` did check for duplicate keys and should have ignored those via the `isIntegrityConstraintViolation(SQLException)` check, but that approach was wrong, because a failure marks the database transaction as "aborted", so having an `ON CONFLICT DO NOTHING` is the right approach. That change simplifies the code a little.

Fixes #7408